Update client to use https too
Client needs to contact https endpoint.

Would be nice to later have a command-line option for http vs https
when running the server.

Note certs with CN only are deprecated (as urllib3 will helpfully tell us)
and have apparently been deprecated for >10 years...
We need a SAN claim in the cert instead, but
`openssl x509` strips it from the csr for security reasons;
we'll get there eventually.